requirements.txt
deepface==0.0.79
gym==0.17.3
gym_super_mario_bros==7.3.0
matplotlib==3.5.2
nes_py==8.2.1
numpy==1.22.1
opencv_python==4.7.0.72
pandas==1.4.2
pandas_datareader==0.10.0
Pillow==9.5.0
plotly==5.9.0
python_dateutil==2.8.2
requests==2.28.1
streamlit==1.21.0
streamlit_chat==0.0.2.2
transformers==4.27.4
yahoo_fin==0.8.9.1
yfinance==0.1.73
